DTREX-406 Refatora API de feature_store
