+++ amora/config.py
@@ -1,12 +1,11 @@
 import logging
 import os
-from enum import Enum
 from uuid import uuid4
 from pathlib import Path
 from tempfile import NamedTemporaryFile
-from typing import Tuple, Dict, Optional
+from typing import Tuple
 
-from pydantic import BaseSettings, AnyUrl
+from pydantic import BaseSettings
 
 ROOT_PATH = Path(__file__).parent.parent
 AMORA_MODULE_PATH = ROOT_PATH.joinpath("amora")
@@ -43,40 +42,4 @@ class Settings(BaseSettings):
         env_prefix = "AMORA_"
 
 
-class FeatureStoreProviders(str, Enum):
-    local = "local"
-    gcp = "gcp"
-
-
-class FeatureStoreOnlineStoreTypes(str, Enum):
-    redis = "redis"
-    sqlite = "sqlite"
-    datastore = "datastore"
-
-
-class FeatureStoreOfflineStoreTypes(str, Enum):
-    bigquery = "bigquery"
-    file = "file"
-
-
-class FeatureStoreSettings(BaseSettings):
-    REGISTRY: str = NamedTemporaryFile(
-        suffix="amora-feature-store-registry", delete=False
-    ).name
-    REPO_PATH: str = NamedTemporaryFile(suffix="repo-path", delete=False).name
-    PROVIDER: str = FeatureStoreProviders.local.value
-    OFFLINE_STORE_TYPE: str = FeatureStoreOfflineStoreTypes.file.value
-    OFFLINE_STORE_CONFIG: Dict[str, str] = {}
-
-    ONLINE_STORE_TYPE: str = FeatureStoreOnlineStoreTypes.sqlite.value
-    ONLINE_STORE_CONFIG: Dict[str, str] = {
-        "path": Path(ROOT_PATH).joinpath("amora-online-feature-store.db").name
-    }
-    DEFAULT_FEATURE_TTL_IN_SECONDS: int = 3600
-
-    class Config:
-        env_prefix = "AMORA_FEATURE_STORE_"
-
-
-feature_store = FeatureStoreSettings()
 settings = Settings()
