+++ amora/feature_store/registry.py
@@ -1,0 +1,7 @@
+from typing import Dict
+
+from feast import FeatureView
+
+from amora.models import Model
+
+FEATURE_REGISTRY: Dict[Model, FeatureView] = {}
